+++ appfair.xcconfig
@@ -10,7 +10,7 @@ PRODUCT_BUNDLE_IDENTIFIER = app.Sun-Bow
 MARKETING_VERSION = 0.0.28
 
 // The monotonically increasing number specifying the internal app version
-CURRENT_PROJECT_VERSION = 187
+CURRENT_PROJECT_VERSION = 188
 
 // The category of the app should be one of: public.app-category.business, developer-tools, education, entertainment, finance, games, graphics-design, healthcare-fitness, lifestyle, medical, music, news, photography, productivity, reference, social-networking, sports, travel, utilities, video, weather
 APP_CATEGORY = public.app-category.utilities
